+++ src/app.js
@@ -1,6 +1,6 @@
 import 'dotenv/config';
 import express from 'express';
-import rankingRouter from './routes/rankingRouter.js';
+import rankingRouter from './routes/ranking.router.js';
 import cookieParser from 'cookie-parser';
 import dotEnv from 'dotenv';
 import errorHandlerMiddleware from './middlewares/error-handler.middleware.js';
@@ -14,12 +14,10 @@ const app = express();
 app.use(express.json());
 app.use(cookieParser());
 app.use('/api', [gamePlayRouter, userRouter, owningPlayerRouter, teamRouter]);
-app.use(errorHandlerMiddleware);
-
 app.use('/api/rankings', rankingRouter);
+app.use(errorHandlerMiddleware);
 
 const PORT = process.env.PORT || 3000;
 app.listen(PORT, () => {
   console.log(`${PORT} 포트로 서버가 열렸어요!`);
 });
-
